+++ src/browser/BrowserPage.cpp
@@ -258,7 +258,7 @@ QImage BrowserPage::getImage(int width, int height, Qt::AspectRatioMode mode)
     }
 
     QSizeF size = m_page->sizeF().scaled(static_cast<int>(width * dApp->devicePixelRatio()), static_cast<int>(height * dApp->devicePixelRatio()), mode);
-    QImage image = m_page->render(size.width(), size.height(), mode);
+    QImage image = m_page->render(static_cast<int>(size.width()), static_cast<int>(size.height()), mode);
     image.setDevicePixelRatio(dApp->devicePixelRatio());
     return image;
 }
@@ -276,7 +276,7 @@ QImage BrowserPage::getImageRect(double scaleFactor, QRect rect)
 QImage BrowserPage::getImagePoint(double scaleFactor, QPoint point)
 {
     QMutexLocker locker(&m_imageMutex);
-    int ss = 122 * scaleFactor / m_scaleFactor;
+    int ss = static_cast<int>(122 * scaleFactor / m_scaleFactor);
 
     QRect rect = QRect(qRound(point.x() * scaleFactor / m_scaleFactor - ss / 2.0), qRound(point.y() * scaleFactor / m_scaleFactor - ss / 2.0), ss, ss);
     return m_page->render(m_rotation, scaleFactor, rect);
